fix: remove remaining builtin validations
