--- packages/validator/src/plugins/validation/2and3/semantic-validators/operations-shared.js
@@ -1,47 +1,0 @@
-//
-// Assertions:
-// 1. An operation should not be defined with a $ref property.
-//
-const pick = require('lodash/pick');
-const map = require('lodash/map');
-const each = require('lodash/each');
-const { hasRefProperty } = require('../../../utils');
-const MessageCarrier = require('../../../utils/message-carrier');
-
-module.exports.validate = function({ jsSpec, resolvedSpec }) {
-  const messages = new MessageCarrier();
-
-  map(resolvedSpec.paths, (path, pathKey) => {
-    if (pathKey.slice(0, 2) === 'x-') {
-      return;
-    }
-    const pathOps = pick(path, [
-      'get',
-      'head',
-      'post',
-      'put',
-      'patch',
-      'delete',
-      'options',
-      'trace'
-    ]);
-
-    each(pathOps, (op, opKey) => {
-      if (!op || op['x-sdk-exclude'] === true) {
-        return;
-      }
-
-      // check for operations that have a $ref property
-      // these are illegal in the spec
-      if (hasRefProperty(jsSpec, ['paths', pathKey, opKey])) {
-        messages.addMessage(
-          `paths.${pathKey}.${opKey}.$ref`,
-          '$ref found in illegal location',
-          'error'
-        );
-      }
-    });
-  });
-
-  return messages;
-};
