--- packages/validator/src/plugins/validation/2and3/structural-validation/validator.js
@@ -1,27 +1,0 @@
-const JSONSchema = require('jsonschema');
-const { transformPathToArray } = require('../../../../path-translator.js');
-
-const jsonSchema = require('./json-schema');
-const { getLineNumberForPath } = require('../../../ast/ast');
-
-const validator = new JSONSchema.Validator();
-validator.addSchema(jsonSchema);
-
-module.exports.validate = function({ jsSpec, specStr, settings = {} }) {
-  settings.schemas.forEach(schema => validator.addSchema(schema));
-  return validator
-    .validate(jsSpec, settings.testSchema || {})
-    .errors.map(err => {
-      return {
-        level: 'error',
-        line: getLineNumberForPath(
-          specStr,
-          transformPathToArray(err.property, jsSpec) || []
-        ),
-        path: err.property.replace('instance.', ''),
-        message: err.message,
-        source: 'schema',
-        original: err // this won't make it into state, but is still helpful
-      };
-    });
-};
